Refactor try_purchase to handle checking all restrictions
Originally we were just passing the results of checking all of the restrictions
into try_purchase, and having it return true, or raising based on that boolean.
The process of checking the restrictions is better handled by try_purchase
itself, since it can just iterate over its own restrictions, and raise when
appropriate.

While we are refactoring try_purchase, we also give a default set of
restrictions to all items (empty set), and change the restrictions to be
classes which we instantiate when checking them to better facilitate testing.